+++ src/Api/RootPackageExpectedException.php
@@ -25,22 +25,38 @@ use RuntimeException;
 class RootPackageExpectedException extends RuntimeException
 {
     /**
+     * Creates an exception for a path mapping that could not be removed from
+     * a non-root package.
+     *
+     * @param string    $repositoryPath The mapped repository path.
+     * @param Exception $cause          The exception that caused this exception.
+     *
+     * @return static The created exception.
+     */
+    public static function cannotRemovePathMapping($repositoryPath, Exception $cause = null)
+    {
+        return new static(sprintf(
+            'Cannot remove path mapping for "%s": Can only remove path '.
+            'mappings from the root package.',
+            $repositoryPath
+        ), 0, $cause);
+    }
+
+    /**
      * Creates an exception for a binding type that could not be removed from
      * a non-root package.
      *
      * @param string    $typeName    The name of the binding type.
-     * @param string    $packageName The name of the package.
      * @param Exception $cause       The exception that caused this exception.
      *
      * @return static The created exception.
      */
-    public static function cannotRemoveBindingType($typeName, $packageName, Exception $cause = null)
+    public static function cannotRemoveBindingType($typeName, Exception $cause = null)
     {
         return new static(sprintf(
-            'Cannot remove binding type "%s" from package "%s": Can only '.
-            'remove types from the root package.',
-            $typeName,
-            $packageName
+            'Cannot remove binding type "%s": Can only remove types from the'.
+            'root package.',
+            $typeName
         ), 0, $cause);
     }